pyproject.toml
[tool.poetry]
name = "team6_webserver"
version = "0.1.0"
description = ""
authors = ["GrolarKim <audwnswmf@naver.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyMySQL = "^1.0.2"
mysql = "^0.0.3"
boto3 = "^1.21.4"
Django = "^4.0.2"
mypy = "^0.931"
django-stubs = "^1.9.0"
types-PyMySQL = "^1.0.13"
Pillow = "^9.0.1"
django-storages = "^1.12.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.mypy]
plugins = ["mypy_django_plugin.main"]
python_version = 3.9
strict = true

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings"

[tool.isort]
profile = "black"
